Move editor to correct filepath
