+++ src/renderer/components/content/editor/aceConfig.js
