+++ src/renderer/components/content/editor/editorView.js
